Change duration to double and pass id to "howl.duration(id)".

+++ src/Howler.Blazor/Components/Howl.Interop.cs
@@ -7,7 +7,7 @@ namespace Howler.Blazor.Components
     public partial class Howl
     {
         [JSInvokable]
-        public void OnPlayCallback(int soundId, int durationInSeconds)
+        public void OnPlayCallback(int soundId, double durationInSeconds)
         {
             TotalTime = TimeSpan.FromSeconds(durationInSeconds);
